requirements.txt
flask
web3
solcx
python-dotenv
pandas
schedule
requests
